refactor and add emitter example to game
