+++ lib/cycle.js
@@ -1,10 +1,5 @@
-var EventEmitter = require('events');
-
-var e = new EventEmitter();
-
-class Cycle extends EventEmitter {
+class Cycle {
   constructor(x, y, number, color, direction) {
-    super();
     this.x         = x;
     this.y         = y;
     this.number    = number;
@@ -42,7 +37,6 @@ class Cycle extends EventEmitter {
     } else {
       this.x -= this.speed;
     }
-    e.emit('someEvent', 'hello');
   }
 
   moveRight() {
@@ -69,15 +63,6 @@ class Cycle extends EventEmitter {
     }
   }
 
-}
-
-e.on('someEvent', function () {
-  console.log('A "someEvent" event was fired.')
-  hello();
-});
-
-function hello() {
-  console.log('hello was called')
 }
 
 module.exports = Cycle;
